+++ location-sample/src/androidTest/java/ru/surfstudio/android/location/sample/mock/TestCustomApp.kt
@@ -3,7 +3,7 @@ package ru.surfstudio.android.location.sample.mock
 import android.app.Application
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
-import ru.surfstudio.android.core.app.ActiveActivityHolder
+import ru.surfstudio.android.activity.holder.ActiveActivityHolder
 import ru.surfstudio.android.location.sample.app.CustomApp
 import ru.surfstudio.android.sample.dagger.app.dagger.DefaultAppModule