+++ RELEASE_NOTES.md
@@ -4,6 +4,11 @@
 
 ## 0.4.0 - SNAPSHOT
 
+#### activity-holder
+
+* Отдельный модуль для `ActiveActivityHolder`
+
+
 #### analytics
 * ANDDEP-294 Доработан модуль аналитики
 * Поддержка любых параметров события аналитики. Возможность отправлять событие только в некоторые аналитики или по условиям.
@@ -15,7 +20,16 @@
 #### broadcast-extension
 #### connection
 #### converter-gson
+
 #### core-app
+
+* Модуль - **deprecated**
+* `CoreApp` - deprecated, при необходимости создайте на конкретном проекте.
+* `ActiveActivityHolder` перенесен в модуль activity-holder
+* `DefaultActivityLifecycleCallbacks` deprecated, реализация по-умолчанию находится в template-модуле
+* `StringsProvider` deprecated, реализация по-умолчанию находится в template-модуле
+* `Unit` - deprecated, используйте kotlin.Unit.
+
 #### core-mvp
 * ANDDEP-320 Возможность динамически задавать LoadState, гибкий способ отображения LoadState
 #### core-mvp-binding
@@ -71,6 +85,9 @@
 #### push
 
 * удален NotificationCenter - используйте `PushHandler`
+* добавлена группировка пуш уведомлений
+* теперь можно подписаться на такие события как открытия и отклонения пуш уведомления
+* по умолчанию при клике на пуш уведомления вызывается колбек `PushEventListener`, чтобы поменять поведение можно переопределить preparePendingIntent у `PushHandlerStrategy`
 
 #### recycler-extension
 
@@ -106,6 +123,14 @@
 * Добавлен плагин Build scans https://guides.gradle.org/creating-build-scans/
 * Добавлена возможность добавить задержку выполнения запроса
     * На экране DebugScreen в настройках сервера можно добавить задержку запроса 0c 0.5c 1c 2c 4c 8c
+* ANDDEP-444 Вынесены Dagger-зависимостей из [`AppComponent`](template/base_feature/src/main/java/ru/surfstudio/standard/application/app/di/AppComponent.kt)
+и [`ActivityComponent`](template/base_feature/src/main/java/ru/surfstudio/standard/ui/activity/di/ActivityComponent.kt)
+в отдельные классы:  [`AppProxyDependencies`](template/base_feature/src/main/java/ru/surfstudio/standard/application/app/di/AppProxyDependencies.kt)
+и [`ActivityProxyDependencies`](template/base_feature/src/main/java/ru/surfstudio/standard/ui/activity/di/ActivityProxyDependencies.kt),
+которые теперь отвечают за распределение зависимостей между компонентами.
+
+
+
 
 #### util-ktx
 
@@ -124,6 +149,13 @@
     * Ускорение работы, добавление проверок на опциональное использование переменных
     * Добавление функций-расширений для опциональной работы с RequestBuilder [`ImageLoaderUtils.kt`](imageloader/src/main/java/ru/surfstudio/android/imageloader/util/ImageLoaderUtils.kt)
     * Добавление расширенного списка стратегий кеширования [`CacheStrategy`](imageloader/src/main/java/ru/surfstudio/android/imageloader/data/CacheStrategy.kt)
+* ANDDEP-442 Расширение функционала [`ImageLoader`](imageloader/src/main/java/ru/surfstudio/android/imageloader/ImageLoader.kt): 
+    * Добавлена поддержка устанавливать Tile изображению (мостить по горизонтали и вертикали)
+    * Изменен метод `ImageLoader.mask`, теперь он принимает параметром `PorterDuff.Mode` для установки необходимого типа заливки
+    * Добавление списка источников загрузки изображения [`ImageSource`](imageloader/src/main/java/ru/surfstudio/android/imageloader/data/ImageSource.kt)
+    * Добавление listener'a с источником загрузки изображения: `ImageLoader.listenerWithSource`
+    * Исправление неочевидного поведения функции `ImageLoader.into`, добавление перегрузки с лямбдами-слушателями
+    
 
 ### security
 * ANDDEP-82 Модуль для обеспечения безопасности